Move to separate file, add documentation
